requirements.txt
# Liste des packages à installer		
category_encoders==1.2.8
eli5==0.8
numpy>=1.15.0
pycebox==0.0.1 
pandas-profiling
PDPbox==0.2.0	
plotly==2.5.1
seaborn
scipy>=1.1.0
shap==0.21.0
skope-rules
